+++ classify/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class AppSvigConfig(AppConfig):
-    name = "svig"
+    name = "classify"
