+++ classify/templates/classify/ajax/category_summary.html
@@ -1,4 +1,4 @@
-{% load svig_tags %}
+{% load custom_tags %}
 
 {% for c in applied_codes %}
 <span class="badge badge-pill badge-{{ c|colour_by_class }}">{{ c }}</span>
