+++ classify/templates/classify/all_classifications.html
@@ -6,7 +6,7 @@
   <div class="container">
     <ol class="breadcrumb" style="background-color: transparent;">
       <li class="breadcrumb-item active"><a href="{% url 'home' %}"><span class="fa fa-home"></span></a></li>
-      <li class="breadcrumb-item"><a href="{% url 'view-all-svig' %}">S-VIG</a></li>
+      <li class="breadcrumb-item"><a href="{% url 'view-all-classifications' %}">Classify</a></li>
     </ol>
   </div>
 </div>
@@ -14,7 +14,7 @@
 
 {% block content %}
 <div class="container">
-    <h5>S-VIG</h5>
+    <h5>Variant classification</h5>
     <br>
     <table class="table" id="classifications-table">
       <thead>
@@ -32,7 +32,7 @@
           <td>{{ c.guideline }}</td>
           <td>{{ c.get_status }}</td>
           <td style="text-align:right">
-            <a class="btn btn-secondary" href="{% url 'svig-analysis' c.pk %}" role="button">View</a>
+            <a class="btn btn-secondary" href="{% url 'perform-classification' c.pk %}" role="button">View</a>
           </td>
         </tr>
         {% endfor %}
