+++ classify/templates/classify/classify_clinical.html
@@ -1,4 +1,4 @@
-{% load svig_tags %}
+{% load custom_tags %}
 {% load crispy_forms_tags %}
