+++ classify/templates/classify/classify_finalise.html
@@ -1,4 +1,4 @@
-{% load svig_tags %}
+{% load custom_tags %}
 {% load crispy_forms_tags %}
 
 <!-- if previous section not completed yet, show warning-->
