+++ classify/templates/classify/classify_info.html
@@ -1,5 +1,5 @@
 {% load crispy_forms_tags %}
-{% load svig_tags %}
+{% load custom_tags %}
 
 <!-- forms when check is complete-->
 {% if classification_info.classification_obj.get_status != "Complete" %}
