--- svig/config/svig_v1.0.yaml
@@ -1,120 +1,0 @@
-codes:
-  O2:
-    type: oncogenic
-    category: mode_of_action
-    options: [SA, VS, ST, MO, SU]
-    description: Null variant in a tumour surpressor gene
-    annotations: []
-  O3:
-    type: oncogenic
-    category: population_db
-    options: [MO]
-    description: Absent or very rare in population database (e.g. gnomAD)
-    annotations: [gnomad]
-  O4:
-    type: oncogenic
-    category: population_db
-    options: [ST, MO, SU]
-    description: Enriched in a somatic variant database (compared to prevalence in controls)
-    annotations: []
-  O5:
-    type: oncogenic
-    category: population_db
-    options: [MO, SU]
-    description: Missense at an amino acid where a different missense variant has previously been determined to be oncogenic, where LOF is a know mechanism of oncogenicity
-    annotations: []
-  O6:
-    type: oncogenic
-    category: population_db
-    options: [SU]
-    description: Computational evidence supports a deleterious effect on the gene or gene product (conservation, evolutionary, splicing impact etc)
-    annotations: []
-  O7:
-    type: oncogenic
-    category: population_db
-    options: [ST, MO, SU]
-    description: Located in a mutational hotspot and/or critical and well-established functional domain (e.g. active site of an enzyme), without benign variation
-    annotations: []
-  O8:
-    type: oncogenic
-    category: population_db
-    options: [MO, SU]
-    description: Protein length change due to in-frame deletions/insertions in a non-repeat region or stop-loss variants
-    annotations: []
-  O9:
-    type: oncogenic
-    category: population_db
-    options: [VS, ST, MO, SU]
-    description: Well-established in vitro or in vivo functional studies demonstrating a functionally abnormal result consistent with the mechanism of disease
-    annotations: []
-  O10:
-    type: oncogenic
-    category: population_db
-    options: [MO, SU]
-    description: Tumour phenotype
-    annotations: []
-  B1:
-    type: benign
-    category: population_db
-    options: [SA, VS, ST, SU]
-    description: Variant is present at a high level in a general population database (e.g. gnomAD)
-    annotations: [gnomad]
-  B2:
-    type: benign
-    category: population_db
-    options: [SA]
-    description: Variant does not fit the mode of action of the gene for the cancer under investigation
-    annotations: []
-  B3:
-    type: benign
-    category: population_db
-    options: [SU]
-    description: Computational evidence DOES NOT support a deleterious effect on the gene product (conservation, evolutionary, splicing impact etc)
-    annotations: []
-  B4:
-    type: benign
-    category: population_db
-    options: [SU]
-    description: Synonymous/ intronic variants (outside canonical splice sites)
-    annotations: []
-  B5:
-    type: benign
-    category: population_db
-    options: [SU]
-    description: In-frame deletion in a repetitive region with unknown function
-    annotations: []
-  B6:
-    type: benign
-    category: population_db
-    options: [ST, MO, SU]
-    description: Well-established in vitro or in vivo functional studies demonstrating no damaging effect on protein function or RNA splicing
-    annotations: []
-  B7:
-    type: benign
-    category: population_db
-    options: [MO, SU]
-    description: Tumour phenotype
-    annotations: []
-
-order:
-  Population databases:
-    - O3_B1
-  Mode of action:
-    - O2
-    - B2
-  Cancer databases:
-    - O4
-  Effect on protein:
-    - O5
-    - O7
-    - O8
-    - B4
-    - B5
-  Computational tools:
-    - O6_B3
-  Functional studies:
-    - O9
-    - B6
-  Other:
-    - O10
-    - B7
