--- svig/config/svig_v1.dev.yaml
@@ -1,32 +1,0 @@
-codes:
-  O2:
-    type: oncogenic
-    category: mode_of_action
-    options: [SA, VS, ST, MO, SU]
-    description: Null variant in a tumour surpressor gene
-    annotations: []
-  O3:
-    type: oncogenic
-    category: population_db
-    options: [MO]
-    description: Absent or very rare in population database (e.g. gnomAD)
-    annotations: []
-  B1:
-    type: benign
-    category: mode_of_action
-    options: [SA, ST, SU]
-    description: Not consistent with mode of action
-    annotations: []
-  B4:
-    type: benign
-    category: population_db
-    options: [SA, VS, ST, SU]
-    description: Variant is present at a high level in a general population database (e.g. gnomAD)
-    annotations: []
-
-order:
-  Population databases:
-    - O3_B4
-  Mode of action:
-    - O2
-    - B1
